requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
idna==3.8
iniconfig==2.0.0
numpy==2.0.2
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==3.13
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
